+++ dungeoncrawler/constants.py
@@ -1,0 +1,8 @@
+SAVE_FILE = "savegame.json"
+SCORE_FILE = "scores.json"
+ANNOUNCER_LINES = [
+    "A decisive blow!",
+    "You fight with determination.",
+    "Your skill improves.",
+    "You press the attack.",
+]
